requirements.txt
Pyrogram==1.4.8
PyYAML==6.0
rich==11.2.0
TgCrypto==1.2.3
